several-changes-interface-classes
the interface logic was modified so that the only viewable Stage at the moment is the main menu; still have to handle an event with the buttons so that these can summon the board or any other stage required. Added a temporary image that serves as a placeholder for the final background menu image, and created a respective `.fxml` file for this menu.