view-controller-relation-added-for-game-finalization
created the fxml for the view and a controller class alredy coded to work together as the final window showed in the game, that displays them in order and gives the user an option to close the game.
extra: deleted a test class that is no longer needed in one of the minigame subpackages.